Refactor: Rebrand finance feature to Pocketly
- Renamed `src/app/finance` directory to `src/app/pocketly`
- Renamed `src/app/api/finance` directory to `src/app/api/pocketly`
- Renamed `src/components/finance-tracker` directory to `src/components/pocketly-tracker`
- Updated hardcoded `/finance` URL strings in frontend components (`ChatbotWidget`, `Admin Layout`)
- Updated internal API fetch calls from `/api/finance` to `/api/pocketly`
- Updated Next.js `middleware.js` to protect `/pocketly` routes instead of `/finance`
- Updated all component import paths reflecting the `finance-tracker` to `pocketly-tracker` rename
