+++ middleware.js
@@ -30,8 +30,8 @@ export default withAuth(
        */
       authorized: ({ token, req }) => {
         const path = req.nextUrl.pathname;
-        // Check if user has admin role for admin routes and finance app
-        if (path.startsWith('/admin') || path.startsWith('/finance')) {
+        // Check if user has admin role for admin routes and pocketly app
+        if (path.startsWith('/admin') || path.startsWith('/pocketly')) {
           return token?.role === 'admin';
         }
         return true;
@@ -41,5 +41,5 @@ export default withAuth(
 );
 
 export const config = {
-  matcher: ['/admin/((?!api|_next/static|_next/image|favicon.ico).*)', '/finance/:path*'],
+  matcher: ['/admin/((?!api|_next/static|_next/image|favicon.ico).*)', '/pocketly/:path*'],
 };
