+++ src/components/pocketly-tracker/CategoriesTab.js
