+++ src/components/pocketly-tracker/FinanceChatBlockRenderer.js
