+++ src/components/pocketly-tracker/IconRenderer.js
